requirements.txt
altgraph==0.17.4
blinker==1.8.1
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
pefile==2023.2.7
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.6
pytube==15.0.0
pywin32-ctypes==0.2.2
setuptools==69.5.1
tqdm==4.66.4
waitress==3.0.0
Werkzeug==3.0.3
yt-dlp==2025.3.31
anyio==4.9.0
certifi==2025.4.26
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
python-telegram-bot==22.1
sniffio==1.3.1
typing_extensions==4.13.2